- move many types from dr2types.h to corresponding files
