+++ src_rebuild/Game/C/event.h
@@ -1,6 +1,37 @@
 #ifndef EVENT_H
 #define EVENT_H
 
+enum SpecialCamera
+{
+	SPECIAL_CAMERA_SET = 0,
+	SPECIAL_CAMERA_SET2 = 1,
+	SPECIAL_CAMERA_RESET = 2,
+	SPECIAL_CAMERA_WAIT = 3,
+};
+
+typedef struct _EVENT EVENT;
+
+struct _EVENT
+{
+	VECTOR position;
+	short rotation;
+	short timer;
+	int* data;
+	int* node;
+	short flags;
+	short radius;
+	int model;
+	EVENT* next;
+};
+
+struct EventGlobal
+{
+	int camera;
+	int draw;
+	EVENT** track;
+	EVENT* cameraEvent;
+};
+
 extern EventGlobal events;
 
 extern CELL_OBJECT *EventCop;
@@ -12,11 +43,6 @@ extern void InitEvents(); // 0x0004BBD4
 extern void SetUpEvents(int full); // 0x00046258
 extern VECTOR* TriggerEvent(int i);
 
-extern void InitEventCamera(); // 0x0004BF54
-extern void ResetEventCamera(); // 0x0004C014
-
-extern void SetCamera(EVENT *ev); // 0x00047538
-
 extern void EventCollisions(CAR_DATA *cp, int type); // 0x0004BC50
 
 extern void StepEvents(); // 0x00048A60
@@ -27,13 +53,9 @@ extern void EventCameraOffset(SVECTOR* offset);
 
 extern sdPlane* EventSurface(VECTOR *pos, sdPlane *plane); // 0x0004A688
 
-extern void MakeEventTrackable(EVENT *ev); // 0x0004BD6C
-
 extern void OffsetTarget(VECTOR *target); // 0x0004BD2C
 
 extern void SetSpecialCamera(SpecialCamera type, int change); // 0x0004B29C
-
-extern void ScreenShake(int count, SVECTOR *ang); // 0x0004C280
 
 extern int DetonatorTimer(); // 0x0004B5FC
