+++ src_rebuild/Game/C/pause.h
@@ -1,6 +1,18 @@
 #ifndef PAUSE_H
 #define PAUSE_H
 
+enum EXIT_VALUE
+{
+	MENU_QUIT_NONE = 0,
+	MENU_QUIT_CONTINUE = 1,
+	MENU_QUIT_QUIT = 2,
+	MENU_QUIT_RESTART = 3,
+	MENU_QUIT_DIRECTOR = 4,
+	MENU_QUIT_QUICKREPLAY = 5,
+	MENU_QUIT_BACKMENU = 6,
+	MENU_QUIT_NEXTMISSION = 7,
+};
+
 extern int gShowMap;
 extern int gDrawPauseMenus;
 extern int pauseflag;
