+++ src_rebuild/Game/C/tile.c
@@ -1,10 +1,10 @@
 #include "driver2.h"
-#include "tile.h"
 #include "models.h"
 #include "system.h"
 #include "mission.h"
 #include "draw.h"
 #include "texture.h"
+#include "tile.h"
 
 #include "ASM/rndrasm.h"
