+++ src_rebuild/Game/C/tile.h
@@ -1,6 +1,27 @@
 #ifndef TILE_H
 #define TILE_H
 
+struct MVERTEX
+{
+	short vx;
+	short vy;
+	short vz;
+	union {
+		short val;
+		struct {
+			u_char u0;
+			u_char v0;
+		}s;
+	}uv;
+};
+
+struct VERTEX
+{
+	DVECTOR coord;
+	UV_INFO uv_coord;
+	u_char pad[2];
+};
+
 extern void DrawTILES(PACKED_CELL_OBJECT** tiles, int tile_amount); // 0x00041D7C
 
 extern void Tile1x1(MODEL *model); // 0x00041B10
